Moved widgets common to screens to dedicated folder
