+++ lib/presentation/screens/common/views/alerts/location_service_alert.dart
@@ -4,10 +4,10 @@ import 'package:movna/domain/entities/location_service_status.dart';
 import 'package:movna/jsons.dart';
 import 'package:movna/presentation/blocs/location_service_cubit.dart';
 import 'package:movna/presentation/locale/locales_helper.dart';
-import 'package:movna/presentation/screens/activity/views/alerts/activity_alert_widget.dart';
-import 'package:movna/presentation/screens/activity/views/alerts/alert_transition_widget.dart';
+import 'package:movna/presentation/screens/common/views/alerts/alert_transition_widget.dart';
+import 'package:movna/presentation/screens/common/views/alerts/alert_widget.dart';
+import 'package:movna/presentation/screens/common/widgets/none_widget.dart';
 import 'package:movna/presentation/theme/app_theme.dart';
-import 'package:movna/presentation/widgets/none_widget.dart';
 
 /// Displays an alert to inform the user the location service is disabled.
 ///
@@ -16,10 +16,10 @@ import 'package:movna/presentation/widgets/none_widget.dart';
 /// See [LocationServiceCubit] the cubit responsible for interacting with the
 /// location service.
 ///
-/// See [ActivityAlertWidget] for more information on the alert
+/// See [AlertWidget] for more information on the alert
 /// rendering.
-class ActivityLocationServiceAlert extends StatelessWidget {
-  const ActivityLocationServiceAlert({super.key});
+class LocationServiceAlert extends StatelessWidget {
+  const LocationServiceAlert({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -31,7 +31,7 @@ class ActivityLocationServiceAlert extends StatelessWidget {
         };
         final Widget child;
         if (status == LocationServiceStatus.disabled) {
-          child = ActivityAlertWidget(
+          child = AlertWidget(
             title: LocaleKeys.location.alert.title().translate(context),
             body: LocaleKeys.location.alert.body().translate(context),
             icon: Icons.location_disabled,
